third_party/WebKit/LayoutTests/external/wpt/html/semantics/interactive-elements/the-dialog-element/centering-expected.txt

This is a testharness.js-based test.
PASS horizontal-tb: tall viewport
PASS horizontal-tb: wide viewport
PASS horizontal-tb: square viewport
PASS horizontal-tb: dialog and viewport match
PASS horizontal-tb: dialog bigger than viewport
FAIL vertical-rl: tall viewport assert_equals: expected "10px" but got "0px"
FAIL vertical-lr: tall viewport assert_equals: expected "10px" but got "0px"
PASS vertical-lr: dialog bigger than viewport
FAIL vertical-rl (dialog horizontal-tb): tall viewport assert_equals: expected "10px" but got "0px"
FAIL vertical-lr (dialog horizontal-tb): tall viewport assert_equals: expected "10px" but got "0px"
PASS vertical-lr (dialog horizontal-tb): dialog bigger than viewport
FAIL horizontal-tb (container vertical-rl): tall viewport assert_equals: expected "10px" but got "0px"
PASS vertical-rl (container horizontal-tb): tall viewport
FAIL horizontal-tb (container vertical-rl) (dialog horizontal-tb): tall viewport assert_equals: expected "10px" but got "0px"
PASS vertical-rl (container horizontal-tb) (dialog vertical-rl): tall viewport
Harness: the test ran to completion.
